Apps/Python/app2/bot/requirements.txt
absl-py==0.11.0
aiodns==3.0.0
aiofiles==0.7.0
aiohttp>=3.7.4
aiomysql==0.0.21
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.4.0
astunparse==1.6.3
async-timeout==3.0.1
asyncio==3.4.3
asyncpg==0.24.0
attrs==21.2.0
audioread==2.1.9
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
better-profanity==0.7.0
cachetools==4.2.1
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
comtypes==1.1.10
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
Deprecated==1.2.12
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@fbb8803779373357e274e1540b368365fd9d8074
discord.py==1.7.3
distlib==0.3.2
docopt==0.6.2
done==1.0.4
dsdev-utils==1.3.0
filelock==3.0.12
flatbuffers==1.12
fuzzywuzzy==0.18.0
gast==0.4.0
gearbox==0.2.1
gevent==21.1.2
greenlet==1.1.0
grpcio==1.34.0
gTTS==2.2.1
gunicorn==20.1.0
httplib2==0.19.1
hupper==1.10.3
idna==2.10
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
Js2Py==0.70
kiwisolver==1.3.1
lazyme==0.0.23
librosa==0.8.0
llvmlite==0.36.0
lru-dict==1.1.7
Markdown==3.3.3
markovify==0.9.3
mediapipe==0.8.6
Menus==0.2.0
multidict==5.1.0
mysql-connector-python==8.0.25
numba==0.53.1
oauthlib==3.1.1
opencv-contrib-python==4.5.2.52
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.5
parso==0.8.2
PasteDeploy==2.1.1
pathlib==1.0.1
pickleshare==0.7.5
Pillow==8.3.2
pipenv==2021.5.29
playsound==1.2.2
pooch==1.3.0
prompt-toolkit==3.0.19
protobuf==3.15.1
psutil==5.8.0
psycopg2==2.9.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycares==4.0.0
pycparser==2.20
pydot==1.4.2
Pygments==2.9.0
pyjokes==0.6.0
pyjsparser==2.7.1
pymongo==3.12.0
PyMySQL==0.9.3
PyNaCl==1.4.0
pyparsing==2.4.7
PyPrind==2.11.2
pyserial==3.5
pySmartDL==1.3.4
python-dateutil==2.8.1
python-Levenshtein==0.12.2
pyttsx3==2.90
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa>=4.7
sacremoses==0.0.43
seaborn==0.11.1
six==1.15.0
slugify==0.0.1
SoundFile==0.10.3.post1
soupsieve==2.2
SQLAlchemy==1.4.21
sqlparse==0.4.2
TagScriptEngine==2.3.0
Tempita==0.5.2
termcolor==1.1.0
threadpoolctl==2.1.0
tokenizers==0.10.1
tqdm==4.57.0
traitlets==5.0.5
transformers==4.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
Unidecode==1.2.0
uritemplate==3.0.1
urllib3>=1.26.5
utils==1.0.1
virtualenv==20.4.7
virtualenv-clone==0.5.4
vosk==0.3.30
wavelink==0.9.10
wcwidth==0.2.5
Werkzeug==2.0.1
wikipedia==1.4.0
wrapt==1.12.1
yarl==1.6.3
youtube-dl==2021.6.6
zope.event==4.5.0
zope.interface==5.4.0
